rename generic.rs to property.rs
